requirements.txt
python-telegram-bot==13.7
firebase-admin==5.0.0
requests==2.26.0
numpy==1.21.2
scikit-learn==0.24.2